Use ip's when syncing from production hosts
Use production host from the inventory with a different hostname
`<hostname>.prod` when creating the dump and syncing to avoid dns/Ansible
issues where the dump would be created on the target machine instead of the
production machine. This requires that the production ip's are specified
in the inventory under the .prod hostname with `ansible_host` and possibly
the user with `ansible_user` if running against Vagrant VM where the used user
doesn't match the one used on the production machine.